refactor: useGetCurrentUrl into own file
